refactoring, reorganize some utils functions
