+++ src/components/Forms/SubmitForm.jsx
@@ -5,7 +5,7 @@ import inputState from '../../utils/submission-validation/input-state';
 import ExpensesContext from '../../store/expenses-context';
 import checkSubmit from '../../utils/submission-validation/check-submit';
 import validateSubmit from '../../utils/submission-validation/validate-submit';
-import formatExpense from '../../utils/format-expense';
+import formatExpense from '../../utils/format-helpers/format-expense';
 import inputChange from '../../utils/submission-validation/input-change';
 import FormInputs from './FormInputs';
 import FormButtons from './FormButtons';
