+++ src/components/Main/Chart/ChartBars.jsx
@@ -2,7 +2,7 @@ import React, { useContext } from 'react';
 import { uniqueId } from 'lodash';
 import ExpensesContext from '../../../store/expenses-context';
 import yearHash from '../../../utils/expenses-reducer/year-hash';
-import monthName from '../../../utils/month-name';
+import monthName from '../../../utils/format-helpers/month-name';
 import ChartBar from './ChartBar';
 
 function ChartBars() {
